third_party/blink/web_tests/wpt_internal/fenced_frame/pointer-lock.https.html

<!DOCTYPE html>
<title>Test of pointer lock</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="resources/utils.js"></script>

<body>

<script>
promise_test(async t => {
  const pointer_lock_key = KEYS['pointer-lock'];

  attachFencedFrame('resources/pointer-lock-inner.html');
  const actual_result = await nextValueFromServer(pointer_lock_key);

  assert_equals(
      actual_result, 'Pointer lock failed',
      'Pointer lock in a fenced frame must fail.');
}, 'Pointer lock in a fenced frame must fail');
</script>

</body>
